refactor(agents): put the spec table where it cannot cycle
The rows lived in packages/agent-<id> behind a ./spec subpath, on the theory
that entry points keep the graph acyclic. They do not -- pnpm and turbo resolve
per PACKAGE, so the moment an agent package gains behavior and depends on
sandbox-docker, turbo refuses:

  Circular package dependency detected: agent-example, agent-registry,
  sandbox-core, relay, ctl, sandbox-docker

So data and behavior split, which is the provider split: the rows are a table in
a leaf, the code is a package loaded from a literal-import map. The four
spec-only packages are deleted rather than left as empty shells; they come back
in phase 2 carrying behavior, the only thing they can hold.

This costs community agents nothing, which was the concern. A plugin agent lives
in the user's node_modules, loads through a variable import(), and never enters
the workspace graph -- structurally exempt. It ships its own descriptor and gets
snapshotted, exactly as provider plugins do.

The purity guard moves with the specs and still checks both halves: the imports
AND the manifest, which is what turbo actually reads. Verified it fires on the
realistic regression -- dep declared and imported.
