--- packages/agent-codex/package.json
@@ -1,41 +1,0 @@
-{
-  "name": "@agentbox/agent-codex",
-  "version": "0.0.0",
-  "private": true,
-  "description": "Codex as an AgentBox agent package \u2014 spec (data) and, later, its behavior",
-  "license": "MIT",
-  "type": "module",
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js"
-    },
-    "./spec": {
-      "types": "./dist/spec.d.ts",
-      "import": "./dist/spec.js"
-    }
-  },
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "build": "tsup",
-    "dev": "tsup --watch",
-    "lint": "eslint src",
-    "test": "vitest run --passWithNoTests",
-    "typecheck": "tsc --noEmit",
-    "clean": "rm -rf dist .turbo"
-  },
-  "dependencies": {
-    "@agentbox/config": "workspace:*",
-    "@agentbox/core": "workspace:*"
-  },
-  "devDependencies": {
-    "@types/node": "^22.10.1",
-    "tsup": "^8.3.5",
-    "typescript": "^5.7.2",
-    "vitest": "^2.1.8"
-  }
-}
