--- packages/agent-example/src/index.ts
@@ -1,9 +1,0 @@
-/**
- * The fourth agent as a package. See `./spec` for what it is and why.
- *
- * Behavior arrives in the same phases as the built-ins', and that is the point:
- * whatever this package still cannot supply is exactly what is still hand-wired
- * elsewhere.
- */
-
-export { exampleSpec } from './spec.js';
