+++ packages/agent-registry/package.json
@@ -2,7 +2,7 @@
   "name": "@agentbox/agent-registry",
   "version": "0.0.0",
   "private": true,
-  "description": "The built-in agent specs, aggregated \u2014 the data half of the agent layer",
+  "description": "The built-in agent specs \u2014 the data half of the agent layer, below sandbox-core",
   "license": "MIT",
   "type": "module",
   "main": "./dist/index.js",
@@ -25,10 +25,7 @@
     "clean": "rm -rf dist .turbo"
   },
   "dependencies": {
-    "@agentbox/agent-claude": "workspace:*",
-    "@agentbox/agent-codex": "workspace:*",
-    "@agentbox/agent-example": "workspace:*",
-    "@agentbox/agent-opencode": "workspace:*",
+    "@agentbox/config": "workspace:*",
     "@agentbox/core": "workspace:*"
   },
   "devDependencies": {
