+++ packages/agent-registry/src/specs/claude.ts
