+++ packages/agent-registry/src/specs/codex.ts
