+++ packages/agent-registry/src/specs/example.ts
