+++ packages/agent-registry/src/specs/opencode.ts
